Renamed `project_configs` and `project_select.py`
